fixed video stream model
